--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@nasc/chalk",
-	"version": "2.3.0",
+	"version": "2.3.1",
 	"description": "Terminal string styling done right",
 	"license": "MIT",
 	"repository": "chalk/chalk",