--- a/test/chalk.js
+++ b/test/chalk.js
@@ -6,7 +6,7 @@ require('./_supports-color')(__dirname);
 const m = require('..');
 
 console.log('TERM:', process.env.TERM || '[none]');
-console.log('platform:', process.platform || '[unknown]');
+console.log('Platform:', process.platform || '[unknown]');
 
 test('don\'t add any styling when called as the base function', t => {
 	t.is(m('foo'), 'foo');