--- a/types/index.d.ts
+++ b/types/index.d.ts
@@ -11,12 +11,10 @@ export const enum Level {
 export interface ChalkOptions {
 	enabled?: boolean;
 	level?: Level;
-	wrapper?: Wrapper;
-}
-
-export interface Wrapper {
-	pre: String,
-	post: String
+	wrapper?: {
+		before: String,
+		after: String
+	}
 }
 
 export interface ChalkConstructor {
@@ -37,7 +35,6 @@ export interface Chalk {
 	constructor: ChalkConstructor;
 	enabled: boolean;
 	level: Level;
-	wrapper: Wrapper;
 	rgb(r: number, g: number, b: number): this;
 	hsl(h: number, s: number, l: number): this;
 	hsv(h: number, s: number, v: number): this;