@sindresorhus 2018-02-11
I'm ok with adding this as it's a valid use-case and it would be convoluted to add it manually.

@Qix- Thoughts?