@sindresorhus 2018-02-11
@felipenmoura It should only be an option in the constructor and not be available on the `chalk` global as a property or as an instance property.