@sindresorhus 2018-02-11
Can you run the benchmark before and after this change to make sure it doesn't introduce any performance regressions?