@Qix- 2018-02-11
I don't see the point in this PR - what is it _really_ trying to solve?

This is what `FORCE_COLOR=1` and `cat -v` (macos) / `cat -A` (linux) / (`man cat` for your arch) are for - so you can debug your output.