@felipenmoura 2018-02-13
Hello.
Sure, I will apply the changes and the benchmark tests and let you know.

About the use-case. I have forked chalk and am using it with these changes to build this:

https://github.com/NascHQ/termtools
(It's still in progress, working better on mac)

The point is...if I use chalk to add colours to the output, _PS1_ will use them as expected...but will also count them as "printable". The colour sequences are "unprintable" characters and need to be wrapped between `\[` and `\]`.
So, neither cat nor force-color will prevent that, as I don't want (in this case) to see the characters, but to "cast" them.

If those characters are not seen as "unprintable" by the PS1, they will apply the colours and _look like_ if they had worked, but when the users start typing, their cursor will "go crazy", moving to the wrong positions in the terminal screen (as it uses the number of printable characters to know the cursor position).

I do believe, though, that `cat -v/A` is, indeed, a nice option for debugging.
As for debugging, it would actually offer the option to "tokenize" the colour characters, allowing simpler Regular Expressions for substitutions or counting, for example.

Also useful when we want to truncate a string...if the string has colours, we may show part of that colour code in the result...and also, we would have counted those characters as the "length" of the string.