@felipenmoura 2018-02-16
As for te benchmarks, I'm getting in average this:

**Before:**  

          144,777 op/s » single style
          44,221 op/s » several styles
         260,126 op/s » cached styles
          46,927 op/s » nested styles

**After:**  

         143,500 op/s » single style
          42,878 op/s » several styles
         271,216 op/s » cached styles
          45,659 op/s » nested styles

Also added a few extra benchmarks (demonstrating there might be _some_ difference from when we _do_ use the wrappers, and when we _don't_ use them):

         143,796 op/s » wrapped single style
          42,430 op/s » wrapped several styles
         238,577 op/s » cached wrapped styles
          44,842 op/s » nested wrapped styles