@felipenmoura 2018-02-19
Hm, I don't get it.

The PR failed building because:

```
examples/rainbow.js:31:36
  ✖  31:36  Unexpected await inside a loop.  no-await-in-loop
```

But I haven't changed a thing in rainbow.js
Any idea what might have caused this?