@Qix- 2018-06-11
@sindresorhus I'm pretty :-1: on this as it's a niche use-case not really suited for chalk. This is what your platform's tools are for and a great use of `FORCE_COLOR`. Unprintable characters are a fact of life, not just in `chalk`, so learning how to debug their output is a pretty essential skill to have.

I don't think introducing _more_ code into `chalk` just for this isolated and ephemeral use-case is valid.