@Qix- 2018-09-19
@sindresorhus I'm still of the above opinion. Should we close this?