@felipenmoura 2018-10-22
No problem, let's close this and move on.
Thanks for your time.